fix color names and s/none/NONE/g
According to :h cterm-colors, this plugin uses the lighter versions of
the colors when it is supposed to use the dark versions. This is also
apparent when you compare the output to terminal output of an ANSI file.

Fix the color names in the highlight items to be the Dark versions.

Change all instances of "none" to "NONE", "none" is now an error in
recent versions of gvim due to a security fix, but "NONE" works
correctly.