fix clusters, add default,bold,underline,reverse
Move the ansiConceal item to the beginning of the file so it has the
lowest priority, otherwise the syn clusters do not work. See
:h syn-priority . Also allow ansiConceal and ansiSuppress to handle any
other ANSI codes that are not explicitly matched so that they are
hidden.

Add default foreground and background (codes 39, 49) clusters and
default items to existing clusters.

Add bold and underline regions and bold and underline items to existing
clusters.

Add another huge set of bg syn clusters for a reverse region, both
background and foreground.

TODO: need to write a test to cross-check all the syntax and highlight
item references.