Only adjust 'highlight' when the conceal feature isn't supported
Prior to the conceal feature, the only way to hide the escape sequences
was by overriding the highlighting of SpecialKey, setting it to Ignore.
When conceal is supported, this is now handled automatically by
concealing the specific ANSI escape sequences.  Therefore, there's no
need to change 'highlight'.

This also means that characters like tab, carriage-return, etc. that
aren't covered by AnsiEsc will be displayed to the user, highlighted by
SpecialKey as expected (especially wrt Tab, since the user may have
configured 'listchars' to show that distinctly).

Finally, this fixes compatibility with Neovim.  'highlight' was removed
from Neovim, so the plugin was emitting errors when trying to set that
option.  However, Neovim always has the conceal feature.