Fix MultiElementHandler to work with non-gui terminals that have termguicolors set.
A "MultiElement", aka a color escape sequence like '\e\[38;5;123m',
is used for "True color" terminals and/or GUI.

Setting the vim variable 'termguicolors' means "use gui colors also in the
terminal, because I know it can handle truecolors (24bit)".

In other words, you basically always want to have termguicolors set
if you are at all interested in MultiElement color escapes and are
on a non-gui terminal like for example konsole (that DOES handle
24bit colors).

The way this works (setting 'termguicolors') is that of the syntax
highligting rules the 'gui=<24bit color>' is used, not the
'cterm=<indexed terminal color>' (same for guifg, ctermfg, guibg and
ctermbg).

It is completely safe to generate syntax rules that contain both;
a GUI will use the 'gui*=' arguments, and ignore the 'cterm*='
arguments, while a non-gui terminal will use the 'cterm*=' arguments
and ignore the 'gui*=' argument UNLESS termguicolors is set, of course,
then it act likes a GUI in that regard.

Not that has("gui") is still 0 for non-gui terminals, even if they
have termguicolors set.

The right thing to do therefore is to not set for has("gui") when
dealing with color codes. One should either test if termguicolors
is set, or simpler, just generate both the 'cterm*=' and 'gui*='
argument.